+++ baby_elf.py
@@ -1,0 +1,6 @@
+from enemy import Enemy
+
+class BabyElf(Enemy):
+    def take_damage(self, damage):
+        self.health -= damage
+        print("Cry, how could you hit a baby 😭, you MONSTER! :(")
